Modify step_toward to clearly indicate a move of 0
When it returned self if move was zero, assigning
the result to a variable made it difficult to
parse if the resulting node was the same as the
destination node because it was returning self or
the child was the destination.

Changing the result to 0 when the move is zero allows
for a clear check that node the method was called
on matched the destination put in.